requirements.txt
absl-py==2.3.1
accelerate==1.11.0
aiofiles==24.1.0
annotated-doc==0.0.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
argcomplete==3.6.3
astunparse==1.6.3
blessed==1.23.0
bracex==2.6
Brotli==1.1.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.17
einops==0.8.1
evo==1.33.0
fastapi==0.120.0
ffmpy==0.6.4
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
fvcore==0.1.5.post20221221
gast==0.6.0
google-pasta==0.2.0
gpustat==1.1.1
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hydra-core==1.3.2
hydra-utils==0.3.2
idna==3.11
imageio==2.37.0
iopath==0.1.10
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
libclang==18.1.1
lz4==4.4.4
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
natsort==8.4.0
networkx==3.5
numexpr==2.14.1
numpy==1.26.1
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti-cu12==12.9.79
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.4.1.4
nvidia-cufile-cu12==1.14.1.1
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.9.79
omegaconf==2.3.0
onnxruntime==1.23.2
opencv-python==4.11.0.86
OpenEXR==3.4.2
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
portalocker==3.2.0
protobuf==6.33.0
psutil==7.1.1
pycolmap==3.13.0
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
roma==1.5.4
rosbags==0.11.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.14.2
safehttpx==0.1.6
safetensors==0.6.2
git+https://github.com/jytime/LightGlue.git#egg=lightglue
scipy==1.16.2
seaborn==0.13.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tiktoken==0.12.0
tomlkit==0.13.3
tqdm==4.67.1
trimesh==4.9.0
triton==3.5.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
versioningit==3.3.0

wcmatch==10.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
yacs==0.1.8
zstandard==0.25.0
